tests: Add postive Atomic test file
