+++ tests/positive/pipeline.cpp
@@ -4747,87 +4747,6 @@ TEST_F(VkPositiveLayerTest, FillBufferCmdPoolTransferQueue) {
     cb.end();
 }
 
-TEST_F(VkPositiveLayerTest, ShaderAtomicInt64) {
-    TEST_DESCRIPTION("Test VK_KHR_shader_atomic_int64.");
-    SetTargetApiVersion(VK_API_VERSION_1_1);
-
-    AddRequiredExtensions(VK_KHR_GET_PHYSICAL_DEVICE_PROPERTIES_2_EXTENSION_NAME);
-    AddRequiredExtensions(VK_KHR_SHADER_ATOMIC_INT64_EXTENSION_NAME);
-    ASSERT_NO_FATAL_FAILURE(InitFramework());
-    if (!AreRequiredExtensionsEnabled()) {
-        GTEST_SKIP() << RequiredExtensionsNotSupported() << " not supported";
-    }
-
-    auto atomic_int64_features = LvlInitStruct<VkPhysicalDeviceShaderAtomicInt64Features>();
-    auto features2 = GetPhysicalDeviceFeatures2(atomic_int64_features);
-    if (features2.features.shaderInt64 == VK_FALSE) {
-        GTEST_SKIP() << "shaderInt64 feature not supported";
-    }
-
-    // at least shaderBufferInt64Atomics is guaranteed to be supported
-    if (atomic_int64_features.shaderBufferInt64Atomics == VK_FALSE) {
-        GTEST_SKIP()
-            << "shaderBufferInt64Atomics feature is required for VK_KHR_shader_atomic_int64 but not expose, likely driver bug";
-    }
-
-    ASSERT_NO_FATAL_FAILURE(InitState(nullptr, &features2));
-
-    if (DeviceValidationVersion() < VK_API_VERSION_1_1) {
-        GTEST_SKIP() << "At least Vulkan version 1.1 is required.";
-    }
-
-    std::string cs_base = R"glsl(
-        #version 450
-        #extension GL_EXT_shader_explicit_arithmetic_types_int64 : enable
-        #extension GL_EXT_shader_atomic_int64 : enable
-        #extension GL_KHR_memory_scope_semantics : enable
-        shared uint64_t x;
-        layout(set = 0, binding = 0) buffer ssbo { uint64_t y; };
-        void main() {
-    )glsl";
-
-    // clang-format off
-    // StorageBuffer storage class
-    std::string cs_storage_buffer = cs_base + R"glsl(
-           atomicAdd(y, 1);
-        }
-    )glsl";
-
-    // StorageBuffer storage class using AtomicStore
-    // atomicStore is slightly different than other atomics, so good edge case
-    std::string cs_store = cs_base + R"glsl(
-           atomicStore(y, 1ul, gl_ScopeDevice, gl_StorageSemanticsBuffer, gl_SemanticsRelaxed);
-        }
-    )glsl";
-
-    // Workgroup storage class
-    std::string cs_workgroup = cs_base + R"glsl(
-           atomicAdd(x, 1);
-           barrier();
-           y = x + 1;
-        }
-    )glsl";
-    // clang-format on
-
-    const char *current_shader = nullptr;
-    const auto set_info = [&](CreateComputePipelineHelper &helper) {
-        // Requires SPIR-V 1.3 for SPV_KHR_storage_buffer_storage_class
-        helper.cs_.reset(new VkShaderObj(this, current_shader, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_1));
-        helper.dsl_bindings_ = {{0, VK_DESCRIPTOR_TYPE_STORAGE_BUFFER, 1, VK_SHADER_STAGE_ALL, nullptr}};
-    };
-
-    current_shader = cs_storage_buffer.c_str();
-    CreateComputePipelineHelper::OneshotTest(*this, set_info, kErrorBit);
-
-    current_shader = cs_store.c_str();
-    CreateComputePipelineHelper::OneshotTest(*this, set_info, kErrorBit);
-
-    if (atomic_int64_features.shaderSharedInt64Atomics == VK_TRUE) {
-        current_shader = cs_workgroup.c_str();
-        CreateComputePipelineHelper::OneshotTest(*this, set_info, kErrorBit);
-    }
-}
-
 TEST_F(VkPositiveLayerTest, TopologyAtRasterizer) {
     TEST_DESCRIPTION("Test topology set when creating a pipeline with tessellation and geometry shader.");
